Align with the evm refactor
