+++ .eslintignore
@@ -1,3 +1,3 @@
 src/model/**
 src/abi/**
-src/substrate/types/**
+src/indexer/substrateIndexer/types/**
