+++ src/indexer/evmIndexer/evmTypes.ts
@@ -2,4 +2,8 @@
 The Licensed Work is (c) 2024 Sygma
 SPDX-License-Identifier: LGPL-3.0-only
 */
-export * as sygmaBridge from './sygma-bridge/events'
+
+export enum ContractType {
+  ERC20 = "erc20",
+  FEE_ROUTER = "feeRouter",
+}
