+++ src/indexer/substrateIndexer/types/events.ts
@@ -1,0 +1,5 @@
+/*
+The Licensed Work is (c) 2024 Sygma
+SPDX-License-Identifier: LGPL-3.0-only
+*/
+export * as sygmaBridge from "./sygma-bridge/events";
