+++ src/indexer/types.ts
@@ -3,6 +3,12 @@ The Licensed Work is (c) 2024 Sygma
 SPDX-License-Identifier: LGPL-3.0-only
 */
 
+export type DepositEvent = {
+  destinationDomainID: number;
+  resourceID: string;
+  depositNonce: bigint;
+  data: string;
+};
 export type DecodedDepositLog = {
   id: string;
   blockNumber: number;
@@ -25,6 +31,7 @@ export type DecodedDepositLog = {
 export type DecodedProposalExecutionLog = {
   id: string;
   blockNumber: number;
+  from: string;
   depositNonce: bigint;
   txHash: string;
   timestamp: Date;
