+++ src/main.ts
@@ -2,58 +2,37 @@
 The Licensed Work is (c) 2024 Sygma
 SPDX-License-Identifier: LGPL-3.0-only
 */
-import { Network } from "@buildwithsygma/sygma-sdk-core";
-import { ethers } from "ethers";
 
-import {
-  getDomainConfig,
-  getProcessorConfig,
-  getSharedConfig,
-  getSsmDomainConfig,
-} from "./config";
-import { startEvmProcessing } from "./evm/evmProcessor";
-import { startSubstrateProcessing } from "./substrate/substrateProcessor";
+import { Network } from "@buildwithsygma/sygma-sdk-core";
+
+import { getConfig } from "./indexer/config";
+import { EVMProcessor } from "./indexer/evmIndexer/evmProcessor";
+import { Indexer } from "./indexer/indexer";
+import type { ISubstrateParser } from "./indexer/substrateIndexer/substrateParser";
+import { SubstrateProcessor } from "./indexer/substrateIndexer/substrateProcessor";
 import { logger } from "./utils/logger";
 
 async function startProcessing(): Promise<void> {
-  const processorConfig = getProcessorConfig();
-  const domainConfig = getDomainConfig();
-  const sharedConfig = await getSharedConfig();
-
-  const thisDomain = sharedConfig.domains.find(
-    (domain) => domain.id == domainConfig.domainID,
-  );
-  if (!thisDomain) {
-    throw new Error(
-      `Domain with ID ${domainConfig.domainID} not found in shared configuration`,
-    );
-  }
-
-  switch (domainConfig.domainType) {
+  const config = await getConfig();
+  let processor;
+  switch (config.domain.type) {
     case Network.EVM: {
-      const provider = new ethers.JsonRpcProvider(domainConfig.rpcURL);
-      const substrateRpcUrlConfig = await getSsmDomainConfig(
-        domainConfig.supportedSubstrateRPCs,
-      );
-
-      logger.info("Process initialization completed successfully.");
-      startEvmProcessing(
-        processorConfig,
-        domainConfig,
-        sharedConfig,
-        thisDomain,
-        provider,
-        substrateRpcUrlConfig,
+      processor = new EVMProcessor(
+        config.parser,
+        config.rpcMap.get(config.domain.id)!,
       );
       break;
     }
     case Network.SUBSTRATE: {
-      startSubstrateProcessing(processorConfig, domainConfig, thisDomain);
+      processor = new SubstrateProcessor(
+        config.parser as ISubstrateParser,
+        config.rpcMap.get(config.domain.id)!,
+      );
       break;
     }
-    default:
-      throw new Error("Unsupported domain type");
   }
+  const indexer = new Indexer(processor, config.domain);
+  indexer.startProcessing();
 }
 
 startProcessing()
