--- src/substrate/types/index.ts
@@ -1,7 +1,0 @@
-/*
-The Licensed Work is (c) 2024 Sygma
-SPDX-License-Identifier: LGPL-3.0-only
-*/
-export * as v1250 from './v1250'
-export * as v1260 from './v1260'
-export * as events from './events'
