+++ typegen.json
@@ -1,5 +1,5 @@
 {
-    "outDir": "src/substrate/types",
+    "outDir": "src/indexer/substrateIndexer/types",
     "specVersions": "https://v2.archive.subsquid.io/metadata/phala",
     "pallets": {
       "SygmaBridge": {
